Use current abi,arch,platform triplet as defaults
If the --binaryAbi, --binaryArch or --binaryPlatform options are not
specified, then use the values for the currently running process as
defaults. Thus the user will not have to look up these values if
future builds for which this repository copy is created, is to be
running on the same (or similar) machine.