Leave out local packages from synchronization
If the version starts with the url scheme 'file:' that means that the
package is fetched from the local filesystem relative to the
package-lock that we were passed; it could for instance be a git
submodule. This package will not be fetched through the web server
that we set up anyway, so don't attempt to synchronize it.